Move VN CfgWeapons classes to separate file
